Move `SortKeyCursor` and `RowIndex` into modules, add `sort_key_cursor` test
